notifier: Add a Date header to emails
This commit adds the current date and time to any email sent using
notifier.SendEmail. This enables the recipient's email client to show a
date and time for the message. The time.RFC1123Z layout matches what RFC
5322 (which obsoletes RFC 2822 and RFC 822) demands for the Date header
and is identical to the layout used by the ETH mail system.